+++ manager_hotel/database/migrations/2023_03_11_063204_create_booking_food_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->integer('id', true);
             $table->integer('booking_id');
             $table->integer('food_id');
-            $table->float('price')->nullable();
+            $table->integer('price')->nullable();
             $table->float('amount');
 
             $table->timestamps();
